+++ math/dec_test.go
@@ -1,0 +1,55 @@
+package math
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestDec(t *testing.T) {
+	zero := Dec{}
+	one := NewDecFromInt64(1)
+	two := NewDecFromInt64(2)
+	three := NewDecFromInt64(3)
+	four := NewDecFromInt64(4)
+	five := NewDecFromInt64(5)
+	minusOne := NewDecFromInt64(-1)
+
+	res, err := two.Add(zero)
+	require.NoError(t, err)
+	require.True(t, res.IsEqual(two))
+
+	res, err = five.Sub(two)
+	require.NoError(t, err)
+	require.True(t, res.IsEqual(three))
+
+	res, err = four.Quo(two)
+	require.NoError(t, err)
+	require.True(t, res.IsEqual(two))
+
+	res, err = five.QuoInteger(two)
+	require.NoError(t, err)
+	require.True(t, res.IsEqual(two))
+
+	res, err = five.Rem(two)
+	require.NoError(t, err)
+	require.True(t, res.IsEqual(one))
+
+	x, err := four.Int64()
+	require.NoError(t, err)
+	require.Equal(t, int64(4), x)
+
+	require.Equal(t, "5", five.String())
+
+	require.True(t, zero.IsZero())
+	require.False(t, zero.IsPositive())
+	require.False(t, zero.IsNegative())
+
+	require.False(t, one.IsZero())
+	require.True(t, one.IsPositive())
+	require.False(t, one.IsNegative())
+
+	require.False(t, minusOne.IsZero())
+	require.False(t, minusOne.IsPositive())
+	require.True(t, minusOne.IsNegative())
+}
