+++ math/math_test.go
@@ -3,56 +3,9 @@ package math
 import (
 	"testing"
 
+	"github.com/cockroachdb/apd/v2"
 	"github.com/stretchr/testify/require"
 )
-
-func TestMath(t *testing.T) {
-	zero := Dec{}
-	one := NewDecFromInt64(1)
-	two := NewDecFromInt64(2)
-	three := NewDecFromInt64(3)
-	four := NewDecFromInt64(4)
-	five := NewDecFromInt64(5)
-	minusOne := NewDecFromInt64(-1)
-
-	res, err := two.Add(zero)
-	require.NoError(t, err)
-	require.True(t, res.IsEqual(two))
-
-	res, err = five.Sub(two)
-	require.NoError(t, err)
-	require.True(t, res.IsEqual(three))
-
-	res, err = four.Quo(two)
-	require.NoError(t, err)
-	require.True(t, res.IsEqual(two))
-
-	res, err = five.QuoInteger(two)
-	require.NoError(t, err)
-	require.True(t, res.IsEqual(two))
-
-	res, err = five.Rem(two)
-	require.NoError(t, err)
-	require.True(t, res.IsEqual(one))
-
-	x, err := four.Int64()
-	require.NoError(t, err)
-	require.Equal(t, int64(4), x)
-
-	require.Equal(t, "5", five.String())
-
-	require.True(t, zero.IsZero())
-	require.False(t, zero.IsPositive())
-	require.False(t, zero.IsNegative())
-
-	require.False(t, one.IsZero())
-	require.True(t, one.IsPositive())
-	require.False(t, one.IsNegative())
-
-	require.False(t, minusOne.IsZero())
-	require.False(t, minusOne.IsPositive())
-	require.True(t, minusOne.IsNegative())
-}
 
 func Test_CountDecPlaces(t *testing.T) {
 	tests := []struct {
@@ -69,12 +22,12 @@ func Test_CountDecPlaces(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.x, func(t *testing.T) {
-			x, err := NewDecFromString(tt.x)
+			x, _, err := apd.NewFromString(tt.x)
 			require.NoError(t, err)
-			got := x.NumDecimalPlaces()
+			got := NumDecimalPlaces(x)
 			require.Equal(t, uint32(tt.want), got)
-			x, _ = x.Reduce()
-			got = x.NumDecimalPlaces()
+			x, _ = x.Reduce(x)
+			got = NumDecimalPlaces(x)
 			require.Equal(t, uint32(tt.reduced), got)
 		})
 	}
